# util: optimize hexToBytes